requirements.txt
# Core RAG dependencies
mistralai>=1.0.0
openai>=1.0.0
supabase>=2.0.0

# Vector database and embeddings
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Text processing
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
tiktoken>=0.5.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
opencv-python>=4.8.0
easyocr>=1.7.0

# Web scraping and data
requests>=2.31.0
aiohttp>=3.9.0

# Database and storage
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Reranking
cohere>=4.0.0

# Langchain integration
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-mistralai>=0.0.1
langchain-cohere>=0.0.1
langchain-supabase>=0.0.1
langchain-core>=0.1.0

# Database ORM
prisma>=0.12.0
prisma-client>=0.12.0

# Optional: Advanced features
chromadb>=0.4.0
pinecone-client>=2.2.0
weaviate-client>=3.25.0
